+++ lib/assets_handlers/local_storage/saveable.dart
@@ -1,0 +1,6 @@
+class Saveable {
+  const Saveable({required this.fromJson, required this.toJson});
+
+  final dynamic Function(String) fromJson;
+  final String Function(dynamic) toJson;
+}