+++ lib/custom_widgets/entity_pages/entity_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:realm_roller/assets_handlers/local_storage/local_storage.dart';
 import 'package:realm_roller/custom_widgets/entity_pages/entity_page_background.dart';
 import 'package:realm_roller/custom_widgets/entity_pages/entity_page_card.dart';
 
@@ -9,12 +10,14 @@ class EntityPage extends StatefulWidget {
     required this.title,
     required this.subtitle,
     required this.imagePath,
+    required this.entity,
     this.children,
   });
 
   final String title;
   final String subtitle;
   final String imagePath;
+  final dynamic entity;
   final List<Widget>? children;
 
   @override
@@ -24,6 +27,7 @@ class EntityPage extends StatefulWidget {
 class _EntityPageState extends State<EntityPage> {
   bool isHidden = false;
   late final ScrollController _scrollController;
+  late bool isSaved;
 
   static const _cardGap = 96.0;
   static const _scrollOffsetThreshold = 7.0;
@@ -45,6 +49,8 @@ class _EntityPageState extends State<EntityPage> {
     super.initState();
     _scrollController = ScrollController();
     _scrollController.addListener(_scrollListener);
+
+    isSaved = LocalStorage.isEntitySaved(widget.entity);
   }
 
   @override
@@ -54,11 +60,21 @@ class _EntityPageState extends State<EntityPage> {
     super.dispose();
   }
 
+  void onSave(bool shouldSave) {
+    if (shouldSave) {
+      LocalStorage.saveEntity(widget.entity);
+    } else {
+      LocalStorage.deleteEntity(widget.entity);
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return EntityPageBackground(
       imagePath: widget.imagePath,
       hideButtons: isHidden,
+      isSaved: isSaved,
+      onSave: onSave,
       children: [
         SingleChildScrollView(
           controller: _scrollController,