+++ lib/custom_widgets/main_menu/main_menu.dart
@@ -4,6 +4,10 @@ import 'package:realm_roller/custom_widgets/main_menu/menu_paged.dart';
 import 'package:realm_roller/custom_widgets/route_builder/route_builder.dart';
 import 'package:realm_roller/pages/general/generators_page/generators_page.dart';
 import 'package:realm_roller/pages/general/main_page/main_page.dart';
+import 'package:realm_roller/pages/general/menu_pages/dice_roller/dice_roller.dart';
+import 'package:realm_roller/pages/general/menu_pages/oracle/oracle_page.dart';
+import 'package:realm_roller/pages/general/menu_pages/saved/saved_page.dart';
+import 'package:realm_roller/pages/general/menu_pages/settings/settings_page.dart';
 
 class MainMenu extends StatelessWidget {
   const MainMenu({super.key, this.currentPage});
@@ -20,10 +24,12 @@ class MainMenu extends StatelessWidget {
   void onGenerators(BuildContext context) =>
       navigate(context, const GeneratorsPage());
 
-  void onDiceRoller(BuildContext context) {}
-  void onOracle(BuildContext context) {}
-  void onSaved(BuildContext context) {}
-  void onSettings(BuildContext context) {}
+  void onDiceRoller(BuildContext context) =>
+      navigate(context, const DiceRoller());
+  void onOracle(BuildContext context) => navigate(context, const OraclePage());
+  void onSaved(BuildContext context) => navigate(context, const SavedPage());
+  void onSettings(BuildContext context) =>
+      navigate(context, const SettingsPage());
 
   @override
   Widget build(BuildContext context) {