+++ lib/main.dart
@@ -1,14 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:realm_roller/assets_handlers/local_storage/local_storage.dart';
+import 'package:realm_roller/assets_handlers/route_observer.dart';
 import 'package:realm_roller/pages/general/main_page/main_page.dart';
 import 'package:realm_roller/theme/theme_data.dart';
 
-void main() {
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   SystemChrome.setPreferredOrientations([
     DeviceOrientation.portraitUp,
     DeviceOrientation.portraitDown,
   ]);
+
+  await LocalStorage.init();
 
   runApp(const MyApp());
 }
@@ -22,6 +26,7 @@ class MyApp extends StatelessWidget {
       title: 'Realm Roller',
       theme: themeData,
       home: const MainPage(),
+      navigatorObservers: [routeObserver],
     );
   }
 }