+++ lib/pages/names/names_generation/settlement_name_generation.dart
@@ -6,6 +6,7 @@ import 'package:randpg/string_manipulations.dart';
 import 'package:realm_roller/assets_handlers/image_path_finders.dart';
 import 'package:realm_roller/custom_widgets/generator_pages/generator_page.dart';
 import 'package:realm_roller/custom_widgets/route_builder/route_builder.dart';
+import 'package:realm_roller/extensions/entities/names_data.dart';
 import 'package:realm_roller/pages/names/names_view/names_view.dart';
 
 import '../../../custom_widgets/dropdowns/dropdown.dart';
@@ -28,26 +29,30 @@ class _SettlementNamesGenerationPageState
 
   void onGenerate(BuildContext context) {
     final settlementType = currentSettlement == "Random"
-        ? ListItemGenerator(SettlementManager.activeSettlementTypes).generate()
-        : SettlementManager.getSettlementTypeByType(
-            currentSettlement.toLowerCase());
+        ? ListItemGenerator(const SettlementManager().activeTypes).generate()
+        : const SettlementManager().getType(currentSettlement.toLowerCase());
 
     final race = currentRace == "Random"
-        ? ListItemGenerator([...RaceManager.activeRaces, null]).generate()
+        ? ListItemGenerator([...const RaceManager().activeTypes, null])
+            .generate()
         : currentRace == "None"
             ? null
-            : RaceManager.getRaceByName(currentRace.toLowerCase());
+            : const RaceManager().getType(currentRace.toLowerCase());
 
     final names =
         UniqueGenerator(settlementType.getNameGenerator(race), _numberOfNames)
             .generate();
 
+    final nameData = SettlementNamesData(
+      names: names,
+      imagePath: getSettlementImage(settlementType),
+      description: titled(
+          "${race?.getAdjective() ?? "mixed"} ${settlementType.getSettlementType()} names"),
+    );
+
     Navigator.of(context).push(buildRoute(
       NamesView(
-        imagePath: getSettlementImage(settlementType),
-        subtitle: titled(
-            "${race?.getAdjective() ?? "mixed"} ${settlementType.getSettlementType()} names"),
-        names: names,
+        namesData: nameData,
       ),
     ));
   }
@@ -79,7 +84,8 @@ class _SettlementNamesGenerationPageState
               onChanged: onSettlementChange,
               options: [
                 "Random",
-                ...SettlementManager.activeSettlementTypes
+                ...const SettlementManager()
+                    .activeTypes
                     .map((e) => titledEach(e.getSettlementType()))
               ],
             ),
@@ -92,7 +98,9 @@ class _SettlementNamesGenerationPageState
               options: [
                 "Random",
                 "None",
-                ...RaceManager.activeRaces.map((e) => titledEach(e.getName()))
+                ...const RaceManager()
+                    .activeTypes
+                    .map((e) => titledEach(e.getName()))
               ],
             ),
           ],