+++ lib/theme/text_theme.dart
@@ -7,75 +7,88 @@ final textTheme = TextTheme(
     fontSize: 93,
     fontWeight: FontWeight.w300,
     letterSpacing: -1.5,
+    color: Colors.black,
   ),
   // H2
   displayMedium: GoogleFonts.medievalSharp(
     fontSize: 58,
     fontWeight: FontWeight.w300,
     letterSpacing: -0.5,
+    color: Colors.black,
   ),
   // H3
   displaySmall: GoogleFonts.medievalSharp(
     fontSize: 47,
     fontWeight: FontWeight.w400,
+    color: Colors.black,
   ),
   // H4
   headlineMedium: GoogleFonts.medievalSharp(
     fontSize: 33,
     fontWeight: FontWeight.w400,
     letterSpacing: 0.25,
+    color: Colors.black,
   ),
   // H5
   headlineSmall: GoogleFonts.medievalSharp(
     fontSize: 23,
     fontWeight: FontWeight.w400,
+    color: Colors.black,
   ),
   // H6
   titleLarge: GoogleFonts.medievalSharp(
     fontSize: 19,
     fontWeight: FontWeight.w500,
     letterSpacing: 0.15,
+    color: Colors.black,
   ),
   // S1
   titleMedium: GoogleFonts.medievalSharp(
     fontSize: 16,
     fontWeight: FontWeight.w400,
     letterSpacing: 0.15,
+    color: Colors.black,
   ),
   // S2
   titleSmall: GoogleFonts.medievalSharp(
     fontSize: 14,
     fontWeight: FontWeight.w500,
     letterSpacing: 0.1,
+    color: Colors.black,
   ),
   // B1
   bodyLarge: GoogleFonts.roboto(
     fontSize: 16,
     fontWeight: FontWeight.w400,
     letterSpacing: 0.5,
+    color: Colors.black,
   ),
   // B2
   bodyMedium: GoogleFonts.roboto(
     fontSize: 14,
     fontWeight: FontWeight.w400,
     letterSpacing: 0.25,
+    color: Colors.black,
   ),
   // Button
   labelLarge: GoogleFonts.roboto(
     fontSize: 14,
     fontWeight: FontWeight.w500,
     letterSpacing: 1.25,
+    color: Colors.black,
   ),
   // Caption
   bodySmall: GoogleFonts.roboto(
     fontSize: 12,
     fontWeight: FontWeight.w400,
     letterSpacing: 0.4,
+    color: Colors.black,
   ),
   // Overline
   labelSmall: GoogleFonts.roboto(
     fontSize: 10,
     fontWeight: FontWeight.w400,
     letterSpacing: 1.5,
+    color: Colors.black,
   ),
 );